app/services/groq_translate.py
from openai import OpenAI
import os
from dotenv import load_dotenv

load_dotenv()

# Initialize client with Groq settings
client = OpenAI(
    api_key=os.getenv("GROQ_API_KEY"),
    base_url=os.getenv("GROQ_ENDPOINT")
)

def translate_with_groq(text: str, target_language: str) -> str:
    try:
        response = client.chat.completions.create(
            model="llama3-70b-8192",
            messages=[
                {
                    "role": "system",
                    "content": f"You are a professional translator. Translate the following text to {target_language}. Only provide the translation, no explanations."
                },
                {
                    "role": "user",
                    "content": text
                }
            ],
            temperature=0.3,
            max_tokens=1024
        )
        return response.choices[0].message.content.strip()
    except Exception as e:
        raise Exception(f"Translation error: {str(e)}")
