requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
openai>=1.0.0
gTTS>=2.3.1
pydantic>=1.8.0